+++ components/cascader-panel/nodeContent.tsx
@@ -2,7 +2,7 @@ import { h, computed, defineComponent, inject, PropType } from 'vue';
 import getPrefixCls from '../_util/getPrefixCls';
 import { CASCADER_PANEL_INJECTION_KEY } from './props';
 
-import type { CascaderNode } from './interface'
+import type { CascaderNode } from './getNode';
 
 const prefixCls = getPrefixCls('cascader-node-content');
